ks_node: refactor telegram auth into unified validateTelegramToken entry point
