+++ key_share_node/server/src/auth/index.ts
@@ -2,5 +2,5 @@ export { validateAuth0Token } from "./auth0";
 export { validateDiscordOAuthToken } from "./discord";
 export { validateGithubOAuthToken } from "./github";
 export { validateGoogleOAuthToken } from "./google";
-export { validateTelegramHash } from "./telegram";
+export { validateTelegramToken } from "./telegram";
 export type { OAuthUser, OAuthValidationFail } from "./types";
